그룹 생성, 가입, 삭제

+++ src/main/java/project/maybedo/repository/GroupRepository.java
@@ -3,5 +3,7 @@ package project.maybedo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.maybedo.domain.Group;
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+import java.util.Optional;
+
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 }
