+++ src/main/java/project/maybedo/repository/JoinRepository.java
@@ -3,5 +3,5 @@ package project.maybedo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.maybedo.domain.Join;
 
-public interface JoinRepository extends JpaRepository<Join, Long> {
+public interface JoinRepository extends JpaRepository<Join, Integer> {
 }
